Created dict_new
